+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [0.3.4] - 2021-05-24
+- **[Added]** Added example of the visualization on first page.
 - **[Added]** `behavior_machine.visualization` which has function `visualize_behavior_machine` that can visualize any machine and it's subcomponents.
 - **[Added]** Examples and tests of visualization under the folder `viz_test`
 - **[Fixed]** Machine now only waits for the remaining time instead of always waiting the fixed amount. Also Fixed tests with weird behavior about the waiting.